test(math): port SimpleRunningAverageTest to network.crypta package
